Add files via upload
A shipping company needs to handle cargo delivery most efficiently and profitably. The company needs to automate the cargo-truck assignment process to achieve good and fair use of its trucks. Fortunately, the company manager discovered your programming skills and deep knowledge of different data structures and decided to hire you to develop a program that simulates the operation of the cargo delivery process and calculates some related statistics to help improve the overall process.